Allow ASSIGNEE to cancel the current RUNNING stage
It’s a bit debatable whether to allow cancel the stage while it’s RUNNING.

On one hand, cancle a RUNNING stage in the middle may cause the task run into a inconsistent stage.

On the other hand, because when there is RUNNING stage, we don’t allow to change the task status. So to prevent bug that making the stage RUN forever, we choose the less annoying option to allow user to cancel the stage.