SQL highlighting on description and comment field in none-edit state.
For edit, highlight js is not intended to be used, the viable option is to use a dedicate code editor. However at this point, we don’t want to introduce a fully fledge editor yet. Thus we are using textarea for edit mode and normal div for none edit mode.

And because of this, we have to carefully adjust the margin to avoid flickering when switching between edit mode.