Remove ws1 prefix in the seed data.
We add prefix mainly because to quickly identify if we fetch the wrong workspace. Since we are always using ws1, we can just treat ws1 as empty prefix, which makes it more realistic.

Because we are doing on-premise deployment and limit to a single workspace, so there is little value to have workspace prefix.