Add role to the Principal object
The role info is stored in the RoleMapping object and we are convert its principalId and updaterId to the corresponding principal, updater. Now we do it in another way, to inject role info into the Principal object.

This is because it’s a more common pattern to access role via the principal instead of via the RoleMapping. The only place that we heavily access the principal and updater object via RoleMapping is the Member UI. And with this change, we introduce a RoleMappingUI to arugment the principal/updater info.

Because of this, we also revsese the fetch dependency in the ProvideContext.