Simplify “String” type custom field, use BBTextField; Also unify to a single place to emit update-custom-field event
Before, we use a complicated field with ok/revert button tracking. The logic is quite compliated and hard to understand.

To simplify, we just use BBTextField. Though by doing this, we lose the capability to restore the text. But to keep it simple for now and enhance into BBTextField later if it’s really needed.