Allow assignee and creator to change custom fields.
For now, we allow creator and assignee to update the field any time. This may cause potential issue that the creator might change some of the fields after the assignee starts the work.

In the future, we could provide options to enforce more strict rules e.g. disallow changing a particular field at a particular stage by a particular role.