Put DataSourceMember list into DataSource intead of creating a separate model
This is to simplify the client code since we need to access the member info frequently in order to determine whether a user can access a particular data source.

Note, we still keep create/delete member under the /member path. This has the benefit of separation of concern and can enforce higher policy.