When fetching database, we also fetch instance directly.
This eliminiates the ordering complexity where we only have the instanceId, whilte the instance for that id hasn’t been fetched yet.

The rule of thumb whether to include related objects is this. For now, we will fetch the container object if needed (e.g. for database, it’s OK to fetch its belonged instance, environment). We need to do this because we usually needs the container object context to function propertly, e.g. database needs the instance info to construct the full url.

On the other hand,  we haven’t seen a use case the otherway around (e.g. for instance, fetching all databases since it might contain too much data). Since the container object doesn’t necessility need the containing model (though it could be inefficient due to the 1+N problem).