Make database route top level and make data source router under database
Move /instance/<<instance slug>>/db/<<db slug>>/.. to /db/<<db slug>>. This includes moving both the database and data source route.

The reason is instance is only exposed to DBA and Owner, while database and data source will also be exposed to developer. If we put database and data source under instance, then we run into the paradox of exposing to instance to the developer while preventing them accessing it.

Also use database instead ds for the url shorthand. So for the data source, the url is /db/<<db slug>>/datasource/<<data source slug>>

Unlike db which is a well known shorthand for database. The data source concept itself is not yet well known, so it’s safe to expose the full name. This is still debatable, another option would be to provide both /ds and /datasource, but that might also causes confusion.