Add dataSourceList to the database model.
To show the readonly indicator in the database list side panel, we need to access the data source membership info.

If the data source membership info is not directly returned in the database, then we won’t get it fromt the data source store since it hasn’t been fetched yet.

Add missing