Simplify the database store to have a single databaseListByInstanceId map to store all database info
UI may fetch the database list from different dimension (by user, by environment). In those cases, we will iterate through this map and compute the list on the fly.

By keeping a single map, we avoid caching inconsistency issue. We save it by instance because database belongs to instance and saving this way follows that hierarchy.

If this causes performance issue, we will add caching later (and deal with the consistency).