Introduce message model and use it for Inbox
Message model will be used in a similar way as in the message bus system.

On the other hand, activity model is bound to a particular scenario, like task activity.

This is still debatable, and we may merge them into one model in the furture if that turns out to be more appropriate.