Remove database/databaseListByUserId
Make DatabaseDashboard use the result database/fetchDatabaseListByUser directly, instead of getting from database/databaseListByUserId later.

The problem with the latter approach is that we store the database by instance id instead of userId and in order to construct the database list by user id, we need to consult the database.project.memberList. However, the project.memberList requires a separate fetch