Add username and password directly to the instance
The username and password served as the default connection string to the instance.

With this, we simplify the logic of not requiring a special * datasource.

For now, we won’t expose the option for user to specify other connection info. This means for a particular instance, user won’t be able to create any new data source, and it means the entire data source feature is not available right now.

In the future, creating a new data source will override the username, password and other connection info defined on the instance.