Use mock/dev/package/demo for building frontend respectively
mock - Using mocked backend (miragjs). This is used during development when we just want to build some frontend feature.

dev - Generate the frontend asset under the backend directory, but with the debug version. This is used during development when we want to build features requiring change both frontend and backend. For now, most backend API is still mocked in the frontend, but overtime, those will be implemented in the real golang backend.

package - Generate the frontend asset under the backend directory, using the production version. This is used for producing a release where we bundle both frontend and backend into a single golang binary.

demo - Using mocked backend (miragjs) but build using the production version. This is used for building the demo version.