Set access_token/refresh_token/user cookie if login successfully.
access_token and refresh_token are http only.

user cookie is NOT http only.

We will regenerate the access_token and refresh_token if the access_token is about to expire.

This implementation still has a problem:

We only regenerate the token if the access_token is present in the request. However, the access_token has a shorter expiration time and the browser would auto purge the expired cookie, so we may end up receiving a request which only has a valid refresh token and won’t able to regenerate the access token.

A followup commit will address this