Implement patch /environment/reorder
Remove the generizied BatchUpdate implementation. Instead, use the specific reorder implementation.

The main motivation is the used google-jsonapi package doesn’t support array annotation like []string, whch is also reasonable as according to JSON API spec, the array data should be put in the outsider block. As a result, we need to tweak the API, so just remove the customized BatchUpdate implementation at all and pass an EnvironmentPatch array instead.