GitLab oauth flow to retrieve access token and refresh token
Step 1. Frontend redirects to GitLab to ask for the access.

Step 2. GitLab redirects to the frontend oauth callback with the code

Step 3. Frontend passes the code to backend to let backend fetches the access_token and refresh_token using the code.

Since the access token has the admin access to the GitLab instance, so we must NOT leak it to the frontend.