Remove statement, rollbackStatement in Issue table and just rely on the correponding payload stored in Task
statement, rollbackStatement could be different for different tasks in a single issue. So it’s better to store there.

Though this does have a tradeoff in that for simple case, and TBH most of the time, the statement/rollbackStatement across all tasks for the same issue are the same. We may add it back to Issue table if the convenience it offers outweighs.