Change to store webhook_endpoint_id instead of the whole webhook url
The full webhook url contains the hostname. The hostname could change, which renders the content invalid. Thus we just store the webhook_endpoint_id which is relative to the webhook url and is sufficient to identify the correponding repo.

Note, if our backend host changes, all the existing webhook would still become invalid since that’s the info stored in VCS like GitLab. But it’s less work to just corret that place.