Move Status (“ACTIVE” | “INVITED”) from principal to member table
By design, when we support multi-workspace in the future. Principal is shared across all workspace, while the same principal can have different member for each workspace. Having status in the member table allow each workspace have “ACTIVE“ | “INVITED” status which is more logical since user is “invited” to a workspace.

Also bring back the UI to allow owner to directly add a member to the workspace.