Add config table
The first use case is to store bb.auth.secret used to sign JWT auth token.

The secret is generated with a 32 letter random string upon creating the config.

Previously we expose the secret flag for user to provide the secret via a file, which is more configurable. But this exposes the extra interface and will generate one more file on the disk. So for the moment, we keep this secret info internal.