Rework how to store the access token/refresh token for the VCS integration
Before this change, we store the token for the user setting up the Git provider. This is problematic for several reasons:

1. The token belongs to the GitLab admin, which has the root privilege to the GitLab instance. This could be a deal breaker to give such a token to external systems like Bytebase.

2. The setup assumes that GitLab admin needs to go to Bytebase to add the Git provider, which is also problematic, because it means that admin needs to be granted the Owner role of Bytebase. However, the GitLab admin may not be the same as Bytebase owner, for a medium size engineering group. The team managing code usually is different from the one managing the database.

What we really need is a token which is sufficient to call VCS api to configure the linked repository. So it’s sufficient to store the access token belonged by the user    who is actually linking the project with the repository.