Initial check in for bb, the Bytebase CLI tool
The first supported command is “bb dump” which behaves similar to the mysqldump and pg_dump.

We will add more commands to bb later.

The idea of bb is it can be used as a standalone tool where the entire Bytebase platform is not needed or can’t be used (e.g. non-UI environment). On the other hand, bb will work better together with the console to deliver a cohesive experience.