Returns database list conditionally based on the query pattern
If caller is NOT requesting for a paritcular instance or the caller is a Developer, then we will only return databases belonging to the project where the caller is a member of.

Also from the instance detail view, if the user is a developer, we only display the database belonging to the project which the user is a member of.


Former-commit-id: 1a45732118d32fb79209dc2425a48cff4286a6b1