Rename backupID to backupId to be consistent and also remove *database member from Backup
Database object won't return to the client, and can always be fetched upfront via databaseId. And we don't have use case to pass this object around, so there is no need to have it in the Backup struct.

Having such object in the struct has overhead, because we need to make sure it's always assigned