After fetching a particular issue, we also update the instance/database store with the fetched instance/database
It might be the first time the particular instance/database objects are returned, so that we should also update instance/database store, otherwise, we may get unknown instance/database when navigating to other UI from the issue detail page since other UIs are getting instance/database by id from the store.

An example is if user navigates to an issue and do a rollback, the constructed rollback issue requires the instance/database exist in the store.


Former-commit-id: 1bcfdd60df5a269a1d7e1bc231a23eb67cd5b30b