Add `status` into mysql_migration_schema.
The status value can be "PENDING" or "DONE"

We record the "PENDING" record before the migration and update that record to "DONE" after migration.

We do this 2-phase commit becuase MySQL runs DDL in its own transaction. Thus we can't record the migration history together with the actual migration.

We may also need this in the future to support long running DDL (DDL appled to large table).