* 🌐 i18n(es-ES.json): update translations for consistency The translations for "back" have been updated to use "Volver" instead of "Atrás" for consistency with the rest of the application. Similarly, "Regresar al inicio" has been updated to "Volver al inicio" for consistency.
* ✨ feat(es-ES.json): add new translations for settings and bbkit
This commit adds new translations for the settings and bbkit sections of the application. The settings section now includes translations for the sidebar, profile, members, im-integration, sso, and other settings. The bbkit section includes translations for common elements, attention, and other components.

* 🎨 style(es-ES.json): capitalize "Inicio de Sesión Único (SSO)"
The commit updates the Spanish translation file for the subscription feature. It capitalizes the "Inicio de Sesión Único (SSO)" string to match the correct capitalization of the words.