* 🎨 style(es-ES.json): fix indentation and spacing The commit message is a minor change to the file that only fixes the indentation and spacing of the file.
* 🐛 fix(es-ES.json): fix syntax error in create-db object
The syntax error was caused by a missing comma after the "new-collection-name" property.

* ✨ feat(es-ES.json): add new translations for creating database users and instances
This commit adds new translations for creating database users and instances. It includes translations for creating admin and read-only users, examples for creating users for different database engines, and sentences for different parts of the user creation process. It also includes translations for testing connections, searching for instance names, and finding GCP project and instance IDs.