chore: update on Spanish Translations (#5855)
* feat(es-ES.json): add new translations for apply-change-to-all-pending-tasks and execution-time
The apply-change-to-all-pending-tasks feature allows users to apply a change to all pending tasks or only to the current task. The execution-time translation is added to display the time taken to execute a task.

* chore(es-ES.json): remove unused translation key
feat(es-ES.json): add translation for "waiting-rollout" issue status
feat(es-ES.json): add translation for "custom-roles" in the settings menu
refactor(es-ES.json): remove unused translation key "sync-schema" and add new keys for schema synchronization
The "waiting-rollout" issue status is now translated to Spanish. The settings menu now includes a translation for "custom-roles". The unused translation key "sync-schema" was removed and replaced with new keys for schema synchronization.

* feat(es-ES.json): add new translations for schema change preview
This commit adds new translations for the schema change preview feature. It includes translations for source-database, source-schema, schema-version, target-databases, with-diff, no-diff, and message. Additionally, it adds a new translation for show-schemaless-databases in the tenant section and new translations for the secret feature. Finally, it adds translations for the feature-require-subscription and advise-index sections.

Former-commit-id: 5a15fe91702348140762183e85019fcda0930138